+++ TabletopTweaks/NewComponents/NestedPsudoActivatableAbilities.cs
@@ -1,6 +1,10 @@
-﻿using Kingmaker.Blueprints;
+﻿using Kingmaker;
+using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.Blueprints.Validation;
+using Kingmaker.Items;
+using Kingmaker.Items.Slots;
+using Kingmaker.PubSubSystem;
 using Kingmaker.UnitLogic;
 using Kingmaker.UnitLogic.Abilities;
 using Kingmaker.UnitLogic.Abilities.Blueprints;
@@ -13,7 +17,7 @@ using TabletopTweaks.NewEvents;
 namespace TabletopTweaks.NewComponents {
     [AllowedOn(typeof(BlueprintAbility), false)]
     [TypeId("a9892b3a72f349fe8acb2e7565d18f93")]
-    class NestedPsudoActivatableAbilities : UnitFactComponentDelegate, ISpontaneousConversionHandler {
+    class NestedPsudoActivatableAbilities : UnitFactComponentDelegate, ISpontaneousConversionHandler, IUnitEquipmentHandler {
         public ReferenceArrayProxy<BlueprintAbility, BlueprintAbilityReference> Variants {
             get {
                 return this.m_Variants;
@@ -35,6 +39,11 @@ namespace TabletopTweaks.NewComponents {
                 AbilityData.AddAbilityUnique(ref conversionList, abilityData);
             }
             conversions = conversionList;
+        }
+
+        public void HandleEquipmentSlotUpdated(ItemSlot slot, ItemEntity previousItem) {
+            if (slot.Owner != Owner) { return; }
+            Game.Instance?.RootUiContext?.InGameVM?.StaticPartVM?.ActionBarVM?.OnUnitChanged(Owner);
         }
 
         public BlueprintAbilityReference[] m_Variants;